Move Fight.hs into Chapter3.hs
